adicionar sistema de reviews com validação de nota e mover configuração do banco para variável de ambiente

+++ app/models/review.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from app.database.db import Base
@@ -9,7 +9,7 @@ class Review(Base):
 
     id_ = Column(Integer, primary_key=True, index=True)
     media_id = Column(Integer, ForeignKey("media.id_"), nullable=False)
-    rating = Column(Integer, nullable=False)
+    rating = Column(Float, nullable=False)
     comment = Column(String(255))
 
     media = relationship("Media", back_populates="reviews")
